+++ internal/day02/day02_test.go
@@ -16,7 +16,7 @@ forward 2`))
 
 func TestPart1(t *testing.T) {
 	got := Part1(testInput)
-	want := int64(150)
+	want := 150
 	if got != want {
 		t.Errorf("Part1() = %d, want %d", got, want)
 	}
@@ -24,7 +24,7 @@ func TestPart1(t *testing.T) {
 
 func TestPart2(t *testing.T) {
 	got := Part2(testInput)
-	want := int64(900)
+	want := 900
 	if got != want {
 		t.Errorf("Part2() = %d, want %d", got, want)
 	}
