+++ internal/day03/day03_test.go
@@ -22,7 +22,7 @@ var testInput = utils.ReadLines(strings.NewReader(`00100
 
 func TestPart1(t *testing.T) {
 	got := Part1(testInput)
-	want := int64(198)
+	want := 198
 	if got != want {
 		t.Errorf("Part1() = %d, want %d", got, want)
 	}
@@ -30,7 +30,7 @@ func TestPart1(t *testing.T) {
 
 func TestPart2(t *testing.T) {
 	got := Part2(testInput)
-	want := int64(230)
+	want := 230
 	if got != want {
 		t.Errorf("Part2() = %d, want %d", got, want)
 	}
