+++ internal/day05/day05_test.go
@@ -20,7 +20,7 @@ var testInput = utils.ReadLines(strings.NewReader(`0,9 -> 5,9
 
 func TestPart1(t *testing.T) {
 	got := Part1(testInput)
-	want := int64(5)
+	want := 5
 	if got != want {
 		t.Errorf("Part1() = %d, want %d", got, want)
 	}
@@ -28,7 +28,7 @@ func TestPart1(t *testing.T) {
 
 func TestPart2(t *testing.T) {
 	got := Part2(testInput)
-	want := int64(12)
+	want := 12
 	if got != want {
 		t.Errorf("Part2() = %d, want %d", got, want)
 	}
