+++ internal/day07/day07_test.go
@@ -11,7 +11,7 @@ var testInput = utils.ReadLines(strings.NewReader(`16,1,2,0,4,2,7,1,2,14`))
 
 func TestPart1(t *testing.T) {
 	got := Part1(testInput)
-	want := int64(37)
+	want := 37
 	if got != want {
 		t.Errorf("Part1() = %d, want %d", got, want)
 	}
@@ -19,7 +19,7 @@ func TestPart1(t *testing.T) {
 
 func TestPart2(t *testing.T) {
 	got := Part2(testInput)
-	want := int64(168)
+	want := 168
 	if got != want {
 		t.Errorf("Part2() = %d, want %d", got, want)
 	}
