+++ internal/day08/day08_test.go
@@ -20,7 +20,7 @@ gcafb gcf dcaebfg ecagb gf abcdeg gaef cafbge fdbac fegbdc |fgae cfgab fg bagce`
 
 func TestPart1(t *testing.T) {
 	got := Part1(testInput)
-	want := int64(26)
+	want := 26
 	if got != want {
 		t.Errorf("Part1() = %d, want %d", got, want)
 	}
@@ -28,7 +28,7 @@ func TestPart1(t *testing.T) {
 
 func TestPart2(t *testing.T) {
 	got := Part2(testInput)
-	want := int64(61229)
+	want := 61229
 	if got != want {
 		t.Errorf("Part2() = %d, want %d", got, want)
 	}
