+++ internal/day09/day09_test.go
@@ -15,7 +15,7 @@ var testInput = utils.ReadLines(strings.NewReader(`2199943210
 
 func TestPart1(t *testing.T) {
 	got := Part1(testInput)
-	want := int64(15)
+	want := 15
 	if got != want {
 		t.Errorf("Part1() = %d, want %d", got, want)
 	}
@@ -23,7 +23,7 @@ func TestPart1(t *testing.T) {
 
 func TestPart2(t *testing.T) {
 	got := Part2(testInput)
-	want := int64(1134)
+	want := 1134
 	if got != want {
 		t.Errorf("Part2() = %d, want %d", got, want)
 	}
