+++ internal/day10/day10_test.go
@@ -20,7 +20,7 @@ var testInput = utils.ReadLines(strings.NewReader(`[({(<(())[]>[[{[]{<()<>>
 
 func TestPart1(t *testing.T) {
 	got := Part1(testInput)
-	want := int64(26397)
+	want := 26397
 	if got != want {
 		t.Errorf("Part1() = %d, want %d", got, want)
 	}
@@ -28,7 +28,7 @@ func TestPart1(t *testing.T) {
 
 func TestPart2(t *testing.T) {
 	got := Part2(testInput)
-	want := int64(288957)
+	want := 288957
 	if got != want {
 		t.Errorf("Part2() = %d, want %d", got, want)
 	}
