+++ internal/day11/day11_test.go
@@ -20,7 +20,7 @@ var testInput = utils.ReadLines(strings.NewReader(`5483143223
 
 func TestPart1(t *testing.T) {
 	got := Part1(testInput)
-	want := int64(1656)
+	want := 1656
 	if got != want {
 		t.Errorf("Part1() = %d, want %d", got, want)
 	}
@@ -28,7 +28,7 @@ func TestPart1(t *testing.T) {
 
 func TestPart2(t *testing.T) {
 	got := Part2(testInput)
-	want := int64(195)
+	want := 195
 	if got != want {
 		t.Errorf("Part2() = %d, want %d", got, want)
 	}
